Move snapshot sets cache to metacache package, store `match` property
